Completed Pages
These are the folders for the homepage of dotCODE and also the JavaScript page. I created both of these over the summer. I am most proud of the JavaScript course and is something I want to share. I wanted a clear and concise teaching method that I just didn't see anywhere else. At some point in the future I want to publish this page and get it live because I do feel like it would help a lot of people.